Factor out Branch and Kore.Step.Simplification.Simplify
