+++ kore/src/Kore/Builtin/Map.hs
@@ -63,9 +63,9 @@ import           Kore.Internal.TermLike
 import qualified Kore.Internal.TermLike as TermLike
 import           Kore.Sort
                  ( Sort )
-import           Kore.Step.Simplification.Data as Simplifier
 import           Kore.Step.Simplification.Data as AttemptedAxiom
                  ( AttemptedAxiom (..) )
+import           Kore.Step.Simplification.Simplify as Simplifier
 import           Kore.Syntax.Sentence
                  ( SentenceSort (..) )
 import           Kore.Unification.Unify
