+++ kore/src/Kore/Step/Simplification/AndPredicates.hs
@@ -13,6 +13,8 @@ module Kore.Step.Simplification.AndPredicates
 
 import qualified Data.Foldable as Foldable
 
+import           Branch
+                 ( BranchT )
 import           Kore.Internal.MultiAnd
                  ( MultiAnd )
 import qualified Kore.Internal.MultiAnd as MultiAnd
@@ -25,10 +27,10 @@ import           Kore.Internal.OrPredicate
                  ( OrPredicate )
 import           Kore.Internal.Pattern
                  ( Predicate )
-import           Kore.Step.Simplification.Data
-                 ( BranchT, MonadSimplify, SimplifierVariable )
 import qualified Kore.Step.Simplification.Data as BranchT
                  ( gather )
+import           Kore.Step.Simplification.Simplify
+                 ( MonadSimplify, SimplifierVariable )
 import qualified Kore.Step.Substitution as Substitution
 
 simplifyEvaluatedMultiPredicate
