+++ kore/test/Test/Kore/Step/Simplification/Predicate.hs
@@ -9,6 +9,7 @@ import Test.Tasty.HUnit
 
 import qualified Data.Map as Map
 
+import           Branch
 import qualified Kore.Internal.MultiOr as MultiOr
 import qualified Kore.Internal.OrPattern as OrPattern
 import           Kore.Internal.OrPredicate
@@ -23,10 +24,11 @@ import           Kore.Step.Axiom.EvaluationStrategy
                  ( firstFullEvaluation )
 import qualified Kore.Step.Axiom.Identifier as AxiomIdentifier
                  ( AxiomIdentifier (..) )
-import           Kore.Step.Simplification.Data hiding
-                 ( runSimplifier )
+import           Kore.Step.Simplification.Data
+                 ( Env (..), evalSimplifier )
 import qualified Kore.Step.Simplification.Predicate as PSSimplifier
                  ( create )
+import           Kore.Step.Simplification.Simplify
 import qualified Kore.Unification.Substitution as Substitution
 import           Kore.Variables.UnifiedVariable
                  ( UnifiedVariable (..) )
